+++ example/lib/main.dart
@@ -14,53 +14,102 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: MyHomePage(title: 'Rounded Loading Button Demo'),
+      home: MyHomePage(),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
-
-  final String title;
+  MyHomePage();
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
 }
 
 class _MyHomePageState extends State<MyHomePage> {
+  final RoundedLoadingButtonController _btnController1 =
+      RoundedLoadingButtonController();
 
-  final RoundedLoadingButtonController _btnController = new RoundedLoadingButtonController();
+  final RoundedLoadingButtonController _btnController2 =
+      RoundedLoadingButtonController();
 
-  void _doSomething() async {
+  void _doSomething(RoundedLoadingButtonController controller) async {
     Timer(Duration(seconds: 3), () {
-      _btnController.success();
+      controller.success();
     });
   }
 
   @override
   Widget build(BuildContext context) {
-
     return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.title),
-      ),
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            RoundedLoadingButton(
-              child: Text('Tap me!', style: TextStyle(color: Colors.white)),
-              controller: _btnController,
-              onPressed: _doSomething,
-              width: 200,
-            ),
-            FlatButton(
-              child: Text("Reset"),
-              onPressed: _btnController.reset)
-          ],
+        appBar: AppBar(
+          title: Text('Rounded Loading Button Demo'),
         ),
-      )
-    );
+        body: Center(
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              RoundedLoadingButton(
+                child: Text('Tap me!', style: TextStyle(color: Colors.white)),
+                controller: _btnController1,
+                onPressed: () => _doSomething(_btnController1),
+              ),
+              SizedBox(
+                height: 50,
+              ),
+              RoundedLoadingButton(
+                color: Colors.amber,
+                successColor: Colors.amber,
+                controller: _btnController2,
+                onPressed: () => _doSomething(_btnController2),
+                valueColor: Colors.black,
+                borderRadius: 10,
+                child: Text('''
+Tap me i have a huge text''', style: TextStyle(color: Colors.white)),
+              ),
+              SizedBox(
+                height: 50,
+              ),
+              OutlinedButton(
+                  style: ButtonStyle(
+                    shape: MaterialStateProperty.all(RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(30))),
+                  ),
+                  onPressed: () {
+                    _btnController1.reset();
+                    _btnController2.reset();
+                  },
+                  child: Text('Reset')),
+              SizedBox(
+                height: 20,
+              ),
+              OutlinedButton(
+                style: ButtonStyle(
+                  shape: MaterialStateProperty.all(RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(30))),
+                ),
+                onPressed: () {
+                  _btnController1.error();
+                  _btnController2.error();
+                },
+                child: Text('Error'),
+              ),
+              SizedBox(
+                height: 20,
+              ),
+              OutlinedButton(
+                style: ButtonStyle(
+                  shape: MaterialStateProperty.all(RoundedRectangleBorder(
+                      borderRadius: BorderRadius.circular(30))),
+                ),
+                onPressed: () {
+                  _btnController1.success();
+                  _btnController2.success();
+                },
+                child: Text('Success'),
+              )
+            ],
+          ),
+        ));
   }
 }